+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 1.2.0 (2019-05-26)
+
+##### New Features
+
+* **build:**  read out project name from process.args to for dynamic extension names [closes#10] (77463f37)
+* **extension:**  add simple kpi extension [closes#9] (65457a57)
+
 #### 1.1.1 (2019-05-24)
 
 ##### Bug Fixes